Compile Fixes, small notes, var additions, damage threshold, etc
Damage threshold enabled, allows mobs per wave/warmup to be customized on the reactor itself, allows for a "safe" vs "dangerous" verification, based on GM/mapper's choice.

Edits the debug message for choose_phrase to be more clear about WHICH reactor is throwing the error.

makes a couple vars in reset_reactor more clear as to which is which.

Standardizes comments in setup_spawner.